+++ src/lib.rs
@@ -101,7 +101,9 @@
 //!     // Adds pinned vertex colors using an Iterator
 //!     .with_pinned_vertex_colors([Color::WHITE, Color::BLACK].into_iter())
 //!     // Adds a single pinned vertex color
-//!     .with_pinned_vertex_color(Color::YELLOW);
+//!     .with_pinned_vertex_color(Color::YELLOW)
+//!     // Adds pinned vertex positions
+//!     .with_pinned_vertex_positions(|pos| pos.x > 0.0 && pos.z <= 5.0);
 //! ```
 //!
 //! For more anchoring options, for example to specify a custom entity to pin the vertices to:
@@ -137,7 +139,9 @@
 //!         // Adds pinned vertex colors using an Iterator
 //!         .with_anchored_vertex_colors([Color::WHITE, Color::BLACK].into_iter(), anchor_to_a)
 //!         // Adds a single pinned vertex color
-//!         .with_anchored_vertex_color(Color::YELLOW, anchor_to_self);
+//!         .with_anchored_vertex_color(Color::YELLOW, anchor_to_self)
+//!         // Adds pinned vertex positions
+//!         .with_anchored_vertex_positions(|pos| pos.x > 0.0 && pos.z <= 5.0, anchor_to_self);
 //! }
 //! ```
 //!